# docs/swagger.yaml
definitions:
...
  handler.AuthResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user: {}
    type: object
...
  handler.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  handler.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        type: string
    type: object
...
  handler.RegisterRequest:
    properties:
      address:
        example: 123 Main St, Anytown, USA
        type: string
      email:
        example: user@example.com
        type: string
      first_name:
        example: John
        type: string
      last_name:
        example: Doe
        type: string
      password:
        example: password123
        type: string
      phone:
        example: 555-123-4567
        type: string
      user_type:
        example: "0"
        type: string
    type: object
  models.Inverter:
    properties:
      capacity:
        type: number
      created_at:
        type: string
      id:
        type: integer
      manufacturer:
        type: string
      model:
        type: string
      updated_at:
        type: string
    type: object
  models.Panel:
    properties:
      created_at:
        type: string
      id:
        type: integer
      longside:
        type: number
      manufacturer:
        type: string
      model:
        type: string
      shortside:
        type: number
      updated_at:
        type: string
      wattage:
        type: number
    type: object
...
  models.Storage:
    properties:
      capacity:
        type: number
      created_at:
        type: string
      id:
        type: integer
      manufacturer:
        type: string
      model:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@sunready.com
    name: API Support
  description: API for Sun Ready project
  license:
    name: Sun Ready Private License
    url: INTERNAL
  termsOfService: http://swagger.io/terms/
  title: Sun Ready API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Login user
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with email and password
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /api/hardware/inverters:
    get:
      description: Get a list of all inverters
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Inverter'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all inverters
      tags:
      - Hardware
    post:
      consumes:
      - application/json
      description: Add a new inverter record
      parameters:
      - description: Inverter payload
        in: body
        name: inverter
        required: true
        schema:
          $ref: '#/definitions/models.Inverter'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Inverter'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add a new inverter
      tags:
      - Hardware
  /api/hardware/panels:
    get:
      description: Get a list of all panels
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Panel'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all panels
      tags:
      - Hardware
    post:
      consumes:
      - application/json
      description: Add a new panel record
      parameters:
      - description: Panel payload
        in: body
        name: panel
        required: true
        schema:
          $ref: '#/definitions/models.Panel'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Panel'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add a new panel
      tags:
      - Hardware
  /api/hardware/storages:
    get:
      description: Get a list of all storage units
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Storage'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all storages
      tags:
      - Hardware
    post:
      consumes:
      - application/json
      description: Add a new storage record
      parameters:
      - description: Storage payload
        in: body
        name: storage
        required: true
        schema:
          $ref: '#/definitions/models.Storage'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Storage'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add a new storage unit
      tags:
      - Hardware
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"